requirements.txt
click==8.1.7
colorama==0.4.6
greenlet==3.0.3
importlib-resources==5.13.0
inflection==0.5.1
mypy==1.10.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
PyMySQL==1.1.1
PyYAML==6.0.1
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.12.2
